pyproject.toml
[tool.poetry]
name = "bank-sales-agent"
version = "0.1.0"
description = "Bank sales support multi-agent prototype built on LangGraph"
authors = ["p-oong <lhp7505@naver.com>"]
readme = "README.md"
packages = [{ include = "bank_sales_agent", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
numpy = "^2.1.3"
streamlit = "^1.44.1"
pydantic = "^2.11.3"
python-dotenv = "^1.1.0"
pyyaml = "^6.0.2"
langgraph = "^1.0.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ruff = "^0.11.6"
black = "^25.1.0"
mypy = "^1.15.0"

[tool.poetry.group.optional]
optional = true

[tool.poetry.group.optional.dependencies]
langgraph-checkpoint-sqlite = "^3.0.3"
langgraph-cli = "^0.4.21"
# TODO: add sentence-transformers when retrieval is introduced.
# TODO: add chromadb or faiss-cpu when local vector storage is introduced.
# TODO: add xgboost or lightgbm when learned ranking is introduced.

[tool.poetry.scripts]
bank-sales-agent = "bank_sales_agent.main:run"
bank-sales-agent-ui = "bank_sales_agent.ui.app:run"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
